Migrate to util
This way we only need 1 function to get ownerRef
